requirements.txt
astroid==2.0.4
dj-database-url==0.5.0
Django==2.1.3
django-bootstrap4==0.0.7
django-cors-headers==2.4.0
django-decouple==2.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.3.0
pkg-resources==0.0.0
psycopg2-binary==2.7.5
PyJWT==1.6.4
pylint==2.1.1
python-decouple==3.1
pytz==2018.7
selenium==3.141.0
six==1.11.0
typed-ast==1.1.0
urllib3==1.24
wrapt==1.10.11
